python/requirements.txt
fastapi=0.135.3
python-dotenv==1.2.2
swagger-ui-py==25.7.1